refactor(schemas): migrate Phase 35 endpoint schemas — 7 files, 8 classes (#6042)
Move local Pydantic class definitions from endpoint files into matching
domain schema modules.

- knowledge_rag_feedback.py: RagFeedbackRequest → schemas_knowledge.py
- knowledge_search_scoped.py: ScopedSearchRequest → schemas_knowledge.py
- skills_repos.py: AddRepoRequest → schemas_code.py (RepoType enum imported
  from skills.models)
- startup.py: StartupPhase + StartupMessage → schemas_system.py
- usage.py: UsageRecordRequest → schemas_analytics.py
- validation_dashboard.py: DashboardGenerateRequest → schemas_workflows.py
- vnc_proxy.py: VncProxyStatusResponse → schemas_system.py

Endpoint files now import from the schemas modules; no behavior change.
