+++ autobot-backend/api/vnc_proxy.py
@@ -14,13 +14,12 @@ enabling the agent to see what users are viewing in real-time.
 
 import asyncio
 import logging
-from typing import Optional
 
 import aiohttp
 from fastapi import APIRouter, Depends, HTTPException, WebSocket, WebSocketDisconnect
 from fastapi.responses import Response
-from pydantic import BaseModel
 
+from api.schemas_system import VncProxyStatusResponse
 from auth_middleware import get_current_user
 from autobot_shared.error_boundaries import ErrorCategory, with_error_handling
 from autobot_shared.http_client import get_http_client
@@ -29,15 +28,6 @@ from type_defs.common import Metadata
 
 router = APIRouter()
 logger = logging.getLogger(__name__)
-
-
-class VncProxyStatusResponse(BaseModel):
-    """Response for VNC proxy status check"""
-    vnc_type: str
-    endpoint: str
-    accessible: bool
-    status: Optional[int] = None
-    error: Optional[str] = None
 
 
 async def _send_client_data_to_vnc(data: dict, vnc_ws, vnc_type: str) -> bool:
